HSEARCH-4133 Move EntityReferenceFactory access from the session to the mapping
Because:

1. We don't need any session information to create an entity reference
   from an entity name + ID.
2. Some of the next changes will deal affect the session and it will be
   simpler if the entity reference factory is not there.
