+++ mapper/javabean/src/main/java/org/hibernate/search/mapper/javabean/session/impl/JavaBeanSearchSessionMappingContext.java
@@ -8,10 +8,15 @@ package org.hibernate.search.mapper.javabean.session.impl;
 
 import java.util.Collection;
 
+import org.hibernate.search.engine.backend.common.spi.EntityReferenceFactory;
+import org.hibernate.search.mapper.javabean.common.EntityReference;
 import org.hibernate.search.mapper.javabean.scope.impl.SearchScopeImpl;
 import org.hibernate.search.mapper.pojo.session.spi.PojoSearchSessionMappingContext;
 
 public interface JavaBeanSearchSessionMappingContext extends PojoSearchSessionMappingContext {
+
+	@Override
+	EntityReferenceFactory<EntityReference> entityReferenceFactory();
 
 	<T> SearchScopeImpl<T> createScope(Collection<? extends Class<? extends T>> types);
