+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/massindexing/impl/HibernateOrmMassIndexingMappingContext.java
@@ -9,7 +9,9 @@ package org.hibernate.search.mapper.orm.massindexing.impl;
 import javax.persistence.EntityManager;
 
 import org.hibernate.engine.spi.SessionFactoryImplementor;
+import org.hibernate.search.engine.backend.common.spi.EntityReferenceFactory;
 import org.hibernate.search.engine.reporting.FailureHandler;
+import org.hibernate.search.mapper.orm.common.EntityReference;
 import org.hibernate.search.mapper.orm.scope.impl.HibernateOrmScopeSessionContext;
 import org.hibernate.search.engine.environment.thread.spi.ThreadPoolProvider;
 
@@ -20,6 +22,8 @@ public interface HibernateOrmMassIndexingMappingContext {
 	ThreadPoolProvider threadPoolProvider();
 
 	FailureHandler failureHandler();
+
+	EntityReferenceFactory<EntityReference> entityReferenceFactory();
 
 	HibernateOrmScopeSessionContext sessionContext(EntityManager entityManager);
