+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/massindexing/impl/HibernateOrmMassIndexingSessionContext.java
@@ -7,8 +7,6 @@
 package org.hibernate.search.mapper.orm.massindexing.impl;
 
 import org.hibernate.engine.spi.SessionImplementor;
-import org.hibernate.search.engine.backend.common.spi.EntityReferenceFactory;
-import org.hibernate.search.mapper.orm.common.EntityReference;
 import org.hibernate.search.mapper.pojo.model.spi.PojoRuntimeIntrospector;
 import org.hibernate.search.mapper.pojo.work.spi.PojoIndexer;
 
@@ -19,7 +17,5 @@ public interface HibernateOrmMassIndexingSessionContext {
 	PojoIndexer createIndexer();
 
 	PojoRuntimeIntrospector runtimeIntrospector();
-
-	EntityReferenceFactory<EntityReference> entityReferenceFactory();
 
 }
