+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/session/impl/HibernateOrmSearchSessionMappingContext.java
@@ -9,7 +9,9 @@ package org.hibernate.search.mapper.orm.session.impl;
 import java.util.Collection;
 
 import org.hibernate.engine.spi.SessionImplementor;
+import org.hibernate.search.engine.backend.common.spi.EntityReferenceFactory;
 import org.hibernate.search.engine.reporting.FailureHandler;
+import org.hibernate.search.mapper.orm.common.EntityReference;
 import org.hibernate.search.mapper.orm.loading.impl.LoadingMappingContext;
 import org.hibernate.search.mapper.orm.scope.impl.SearchScopeImpl;
 import org.hibernate.search.mapper.pojo.session.spi.PojoSearchSessionMappingContext;
@@ -18,6 +20,9 @@ public interface HibernateOrmSearchSessionMappingContext
 		extends PojoSearchSessionMappingContext, LoadingMappingContext {
 
 	FailureHandler failureHandler();
+
+	@Override
+	EntityReferenceFactory<EntityReference> entityReferenceFactory();
 
 	<T> SearchScopeImpl<T> createScope(Collection<? extends Class<? extends T>> types);
