+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/work/spi/PojoWorkMappingContext.java
@@ -6,6 +6,7 @@
  */
 package org.hibernate.search.mapper.pojo.work.spi;
 
+import org.hibernate.search.engine.backend.common.spi.EntityReferenceFactory;
 import org.hibernate.search.engine.backend.mapping.spi.BackendMappingContext;
 import org.hibernate.search.mapper.pojo.bridge.runtime.spi.BridgeMappingContext;
 
@@ -13,5 +14,7 @@ import org.hibernate.search.mapper.pojo.bridge.runtime.spi.BridgeMappingContext;
  * Mapping-scoped information and operations for use in POJO work execution.
  */
 public interface PojoWorkMappingContext extends BackendMappingContext, BridgeMappingContext {
+
+	EntityReferenceFactory<?> entityReferenceFactory();
 
 }
