+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/work/spi/PojoWorkSessionContext.java
@@ -6,7 +6,6 @@
  */
 package org.hibernate.search.mapper.pojo.work.spi;
 
-import org.hibernate.search.engine.backend.common.spi.EntityReferenceFactory;
 import org.hibernate.search.engine.backend.session.spi.BackendSessionContext;
 import org.hibernate.search.mapper.pojo.automaticindexing.spi.PojoImplicitReindexingResolverSessionContext;
 import org.hibernate.search.mapper.pojo.bridge.runtime.spi.BridgeSessionContext;
@@ -22,8 +21,6 @@ public interface PojoWorkSessionContext
 
 	@Override
 	PojoWorkMappingContext mappingContext();
-
-	EntityReferenceFactory<?> entityReferenceFactory();
 
 	PojoLoadingContext defaultLoadingContext();
